+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 *Minor*
 
 * Replaced references to Millers Hollow with an original game description. ([#126](https://github.com/hjwylde/werewolf/issues/126))
+* Removed the advice field from Role. ([#134](https://github.com/hjwylde/werewolf/issues/134))
+* Updated the Seer's description. ([#134](https://github.com/hjwylde/werewolf/issues/134))
 
 *Revisions*