+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fixed a bug where extra roles with spaces in them weren't recognised.
 * Fixed Devoted Servant messages to exclude her when joining the Werewolf pack.
 * Fixed Devoted Servant help messages to include how to `pass`.
+* Removed player cap of 24. ([#143](https://github.com/hjwylde/werewolf/issues/143))
 
 ### v0.4.10.0