+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Fixed Devoted Servant messages to exclude her when joining the Werewolf pack.
 * Fixed Devoted Servant help messages to include how to `pass`.
 * Removed player cap of 24. ([#143](https://github.com/hjwylde/werewolf/issues/143))
+* Fixed a bug where the Wild Child would receive role model died messages when dead. ([#145](https://github.com/hjwylde/werewolf/issues/145))
 
 ### v0.4.10.0