+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Updated the Simple Villager's description and rules. ([#135](https://github.com/hjwylde/werewolf/issues/135))
 * Removed the advice field from Role. ([#134](https://github.com/hjwylde/werewolf/issues/134))
 * Updated the Seer's description. ([#134](https://github.com/hjwylde/werewolf/issues/134))
+* Updated the Simple Werewolf's description. ([#139](https://github.com/hjwylde/werewolf/issues/139))
 
 *Revisions*