+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Upcoming
 
+### v0.4.11.0
+
 *Minor*
 
 * Replaced references to Millers Hollow with an original game description. ([#126](https://github.com/hjwylde/werewolf/issues/126))